pyproject.toml
[project]
name = "comfyui_ryanonyheinside"
description = "Custom nodes introducing particle simulations, optical flow, audio manipulation & reactivity, and temporal masks"
version = "2.0.0"
license = {file = "LICENSE"}
dependencies = ["pygame","opencv-python==4.10.0","scipy","torchaudio", "pillow","librosa==0.10.2","pymunk==6.8.1","matplotlib","openunmix","mido","scikit-image"]

[project.urls]
Repository = "https://github.com/ryanontheinside/ComfyUI_RyanOnTheInside"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "ryanontheinside"
DisplayName = "RyanOnTheInside"
Icon = "🌀"
